Move debug_flags definition out of the generator
