+++ ash/src/extensions/experimental/amd.rs
@@ -25,7 +25,9 @@
  *
  **********************************************************************************************************************/
 
+use crate::prelude::debug_flags;
 use crate::vk::*;
+
 use std::fmt;
 use std::os::raw::*;
