+++ ash/src/vk.rs
@@ -11,7 +11,6 @@ pub use aliases::*;
 mod bitflags;
 pub use bitflags::*;
 mod const_debugs;
-pub(crate) use const_debugs::*;
 mod constants;
 pub use constants::*;
 mod definitions;
